Address more Alex feedback
