+++ logfire/__init__.py
@@ -27,7 +27,6 @@ from ._internal.stack_info import add_non_user_code_prefix
 from ._internal.utils import suppress_instrumentation
 from .integrations.logging import LogfireLoggingHandler
 from .integrations.structlog import LogfireProcessor as StructlogProcessor
-from .variables.push import push_variables, validate_variables
 from .version import VERSION
 
 DEFAULT_LOGFIRE_INSTANCE: Logfire = Logfire()
@@ -96,6 +95,8 @@ metric_up_down_counter_callback = DEFAULT_LOGFIRE_INSTANCE.metric_up_down_counte
 # Variables
 var = DEFAULT_LOGFIRE_INSTANCE.var
 get_variables = DEFAULT_LOGFIRE_INSTANCE.get_variables
+push_variables = DEFAULT_LOGFIRE_INSTANCE.push_variables
+validate_variables = DEFAULT_LOGFIRE_INSTANCE.validate_variables
 
 
 def loguru_handler() -> Any:
